rename route component to avoid next convention error. node version package updated to avoid warning
