+++ package.json
@@ -48,6 +48,6 @@
 		"typescript": "5"
 	},
 	"engines": {
-		"node": ">=18.17.0"
+		"node": "18.17.0"
 	}
 }
