+++ src/app/[location]/[place]/[sector]/[boulder]/components/route-component.tsx
